Fix: changed the numeric type of multiple files necessary (float  ->  int)

+++ print_bigger_number_from_two_user_inputs.py
@@ -1,8 +1,8 @@
 # B1 Prog01: Create a program that ask user to input 2 numbers. Print the bigger number.
 
 # User Input
-num1 = float(input("First Number: "))
-num2 = float(input("Second Number: "))
+num1 = int(input("First Number: "))
+num2 = int(input("Second Number: "))
 
 # If-else statement for choosing bigger number
 if num1 > num2:
